refactor(front): move getLargeWhitelistedModel to lib/api/assistant/models
