+++ front/lib/api/assistant/global_agents/configurations/dust/deep-dive.ts
@@ -19,8 +19,10 @@ import {
   _getToolsetsToolsConfiguration,
 } from "@app/lib/api/assistant/global_agents/tools";
 import { dummyModelConfiguration } from "@app/lib/api/assistant/global_agents/utils";
-import { isProviderWhitelisted } from "@app/lib/api/assistant/models";
-import { getLargeWhitelistedModel } from "@app/lib/assistant";
+import {
+  getLargeWhitelistedModel,
+  isProviderWhitelisted,
+} from "@app/lib/api/assistant/models";
 import type { Authenticator } from "@app/lib/auth";
 import type { GlobalAgentSettingsModel } from "@app/lib/models/agent/agent";
 import type {
