+++ front/lib/api/assistant/global_agents/configurations/dust/sidekick.ts
@@ -7,10 +7,10 @@ import type {
 } from "@app/lib/api/assistant/global_agents/tools";
 import { dummyModelConfiguration } from "@app/lib/api/assistant/global_agents/utils";
 import {
+  getLargeWhitelistedModel,
   getSmallWhitelistedModel,
   isProviderWhitelisted,
 } from "@app/lib/api/assistant/models";
-import { getLargeWhitelistedModel } from "@app/lib/assistant";
 import type { Authenticator } from "@app/lib/auth";
 import type {
   AgentConfigurationType,
